# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/files":{"post":{"operationId":"FilesController_create","summary":"File or Image Upload Dynamic Bucket","parameters":[{"name":"Authorization","in":"header","description":"Bearer Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileUploadDTO"}}}},"responses":{"201":{"description":"Return upload information."},"400":{"description":"Invalid data"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Files"],"security":[{"bearer":[]}]},"get":{"operationId":"FilesController_findAll","summary":"Fetches records","parameters":[],"responses":{"200":{"description":"Returns list of records."},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Files"]}},"/files/{id}":{"get":{"operationId":"FilesController_findOne","summary":"Get a record by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns record."},"404":{"description":"Record Not found."},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Files"]},"patch":{"operationId":"FilesController_update","summary":"Update record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFileDto"}}}},"responses":{"200":{"description":"Return record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Files"]},"delete":{"operationId":"FilesController_remove","summary":"delete record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Files"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/login":{"post":{"operationId":"AuthController_login","summary":"User login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDTO"}}}},"responses":{"200":{"description":"Return user information."}},"tags":["Authentication"]}},"/folders":{"post":{"operationId":"FoldersController_create","summary":"Create a new record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFolderDto"}}}},"responses":{"201":{"description":"Return new record."},"400":{"description":"Invalid data"},"405":{"description":"Method not allowed"},"406":{"description":"Record already exist"},"500":{"description":"Server Error!"}},"tags":["Folders"]},"get":{"operationId":"FoldersController_findAll","summary":"Fetches records","parameters":[],"responses":{"200":{"description":"Returns list of records."},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Folders"]}},"/folders/{id}":{"get":{"operationId":"FoldersController_findOne","summary":"Get a record by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns record."},"404":{"description":"Record Not found."},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Folders"]},"patch":{"operationId":"FoldersController_update","summary":"Update record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFolderDto"}}}},"responses":{"200":{"description":"Return record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Folders"]},"delete":{"operationId":"FoldersController_remove","summary":"delete record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Folders"]}}},"info":{"title":"Api","description":"This api will help clients to store their data.","version":"1.0","contact":{}},"tags":[{"name":"Api","description":""}],"servers":[],"components":{"schemas":{"FileUploadDTO":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]},"Folder":{"type":"object","properties":{}},"UpdateFileDto":{"type":"object","properties":{"folder":{"$ref":"#/components/schemas/Folder"}},"required":["folder"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","password","firstName","lastName"]},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"isActive":{"type":"boolean"},"isDeleted":{"type":"boolean"}},"required":["firstName","lastName","isActive","isDeleted"]},"AuthDTO":{"type":"object","properties":{}},"User":{"type":"object","properties":{}},"CreateFolderDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"folder":{"$ref":"#/components/schemas/Folder"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","name","folder","user"]},"UpdateFolderDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"folder":{"$ref":"#/components/schemas/Folder"},"user":{"$ref":"#/components/schemas/User"}}}}}}